+++ src/main/java/org/gradlex/jvm/dependency/conflict/detection/rules/logging/LoggingModuleIdentifiers.java
@@ -23,6 +23,7 @@ public enum LoggingModuleIdentifiers {
     LOG4J_SLF4J2_IMPL("org.apache.logging.log4j", "log4j-slf4j2-impl", "2.19.0"),
     LOG4J_TO_SLF4J("org.apache.logging.log4j", "log4j-to-slf4j", "2.0"),
     SLF4J_SIMPLE("org.slf4j", "slf4j-simple", "1.0"),
+    SLF4J_NOP("org.slf4j", "slf4j-nop", "1.0"),
     LOGBACK_CLASSIC("ch.qos.logback", "logback-classic", "1.0.0"),
     SLF4J_LOG4J12("org.slf4j", "slf4j-log4j12", "1.0"),
     SLF4J_JCL("org.slf4j", "slf4j-jcl", "1.0"),